tests/krate/publish: Extract `build_metadata` module
