+++ src/tests/krate/publish/mod.rs
@@ -1,5 +1,4 @@
 use crate::new_category;
-use crate::util::insta::assert_yaml_snapshot;
 use crate::util::{RequestHelper, TestApp};
 use crate::{
     builders::{CrateBuilder, DependencyBuilder, PublishBuilder},
@@ -24,6 +23,7 @@ use std::iter::FromIterator;
 use std::time::Duration;
 use std::{io, thread};
 
+mod build_metadata;
 mod inheritance;
 
 #[test]
@@ -1320,57 +1320,6 @@ fn empty_payload() {
     );
 
     assert!(app.stored_files().is_empty());
-}
-
-fn version_with_build_metadata(v1: &str, v2: &str, expected_error: &str) {
-    let (_app, _anon, _cookie, token) = TestApp::full().with_token();
-
-    let response = token.publish_crate(PublishBuilder::new("foo", v1));
-    assert_eq!(response.status(), StatusCode::OK);
-    assert_yaml_snapshot!(response.into_json(), {
-        ".crate.created_at" => "[datetime]",
-        ".crate.updated_at" => "[datetime]",
-    });
-
-    let response = token.publish_crate(PublishBuilder::new("foo", v2));
-    assert_eq!(response.status(), StatusCode::OK);
-    assert_eq!(
-        response.into_json(),
-        json!({ "errors": [{ "detail": expected_error }] })
-    );
-}
-
-#[test]
-fn version_with_build_metadata_1() {
-    insta::with_settings!({ snapshot_suffix => "build_metadata_1" }, {
-        version_with_build_metadata(
-            "1.0.0+foo",
-            "1.0.0+bar",
-            "crate version `1.0.0` is already uploaded",
-        );
-    });
-}
-
-#[test]
-fn version_with_build_metadata_2() {
-    insta::with_settings!({ snapshot_suffix => "build_metadata_2" }, {
-        version_with_build_metadata(
-            "1.0.0-beta.1",
-            "1.0.0-beta.1+2",
-            "crate version `1.0.0-beta.1` is already uploaded",
-        );
-    });
-}
-
-#[test]
-fn version_with_build_metadata_3() {
-    insta::with_settings!({ snapshot_suffix => "build_metadata_3" }, {
-        version_with_build_metadata(
-            "1.0.0+foo",
-            "1.0.0",
-            "crate version `1.0.0` is already uploaded",
-        );
-    });
 }
 
 #[test]
